+++ CHANGELOG.md
@@ -36,6 +36,8 @@ PaginationBox
 
 - selectTableHOC SelectTable 移除
 
+- 默认没有top, left, right 的border
+
 ### css
 
 移除 .ifont- .xfont- 等功能性样式，替换为 .gm-svg-icon
@@ -50,3 +52,4 @@ remove src/framework
 
 List 的 data 改为 [{value, text}] ,renderName 改为 renderItem
 remove ListGroup，统统在 List 里
+